chore: storybook 설정 파일 수정
react-router-dom 관련 설정 추가
